Rename userAttributes folder
